requirements.txt
streamlit 
sqlalchemy 
psycopg2-binary 
ultralytics
opencv-python
Pillow
